Move logic from kubernetesDeploy to kubernetes package
