+++ resources/metadata/helmExecute.yaml
@@ -268,6 +268,112 @@ spec:
           - PARAMETERS
           - STAGES
           - STEPS
+      - name: valuesMapping
+        type: "map[string]interface{}"
+        longDescription: |
+          Mapping of values provided by Piper onto custom paths in format `[custom-path]: [piper-value]`
+
+          Example:
+          ```yaml
+          valuesMapping:
+            subchart.image.tag:        image.debug.tag
+            subchart.image.repository: image.debug.repository
+            subchart.image.pullsecret: secret.dockerconfigjson
+          ```
+        scope:
+          - PARAMETERS
+          - STAGES
+          - STEPS
+      - name: imageNames
+        type: "[]string"
+        description: List of names of the images to be deployed.
+        resourceRef:
+          - name: commonPipelineEnvironment
+            param: container/imageNames
+        scope:
+          - PARAMETERS
+          - STAGES
+          - STEPS
+      - name: imageNameTags
+        type: "[]string"
+        description: List of full names (registry and tag) of the images to be deployed.
+        resourceRef:
+          - name: commonPipelineEnvironment
+            param: container/imageNameTags
+        scope:
+          - PARAMETERS
+          - STAGES
+          - STEPS
+      - name: imageDigests
+        type: "[]string"
+        description: List of image digests of the images to be deployed, in the format `sha256:<hash>`. If provided, image digests will be appended to the image tag, e.g. `<repository>/<name>:<tag>@<digest>`
+        resourceRef:
+          - name: commonPipelineEnvironment
+            param: container/imageDigests
+        scope:
+          - PARAMETERS
+          - STAGES
+          - STEPS
+      - name: appTemplate
+        aliases:
+          - name: k8sAppTemplate
+        type: string
+        description: Defines the filename for the kubernetes app template (e.g. k8s_apptemplate.yaml).
+        longDescription: |
+          There are two supported ways for the template rendering:
+
+          1. For a deployments using single image, you can use a placeholder `<image-name>`, which will be replaced with the image GUN.
+          ```
+          apiVersion: apps/v1
+          kind: Deployment
+          metadata:
+            name: app
+            labels:
+              app: app
+          spec:
+            replicas: 3
+            selector:
+              matchLabels:
+                app: app
+            template:
+              metadata:
+                labels:
+                  app: app
+              spec:
+                containers:
+                - name: app
+                  image: <image-name>
+          ```
+
+          2. Helm styled templates, with the support for multi-image deployments.
+
+          ```
+          apiVersion: apps/v1
+          kind: Deployment
+          metadata:
+            name: app
+            labels:
+              app: app
+          spec:
+            replicas: 3
+            selector:
+              matchLabels:
+                app: app
+            template:
+              metadata:
+                labels:
+                  app: app
+              spec:
+                containers:
+                - name: app-1
+                  image: "{{ .Values.image.repository}}:{{ .Values.image.tag }}"
+                - name: app-2
+                  image: "{{ .Values.image.app_2.repository}}:{{ .Values.image.app_2.tag }}"
+          ```
+        scope:
+          - PARAMETERS
+          - STAGES
+          - STEPS
   containers:
     - image: dtzar/helm-kubectl:3.8.0
       workingDir: /config
